Separate the shard code for maven and npm of compute_normalized_license function #1403
